Moving miscellaneous helper functions to utils.py
